requirements.txt
torchaudio
torchvision
plyfile
tensorboard
submodules/diff-gaussian-rasterization
submodules/simple-knn